styled interface, more input fields

+++ src/App.js
@@ -1,34 +1,32 @@
-import './App.css';
+//==== Includes ====
+
+/* React */
+import './App.scss';
 import { useState, useEffect } from 'react';
- 
+
+/* Firebase */ 
 import { addDoc, getDocs, collection, doc, deleteDoc, updateDoc } from "@firebase/firestore"
 import { db } from "./firebase-config"
 
-//==== Form parameters
-const INPUT_NAMES = ["name", "surname", "email", "telephone"];	// Defines properties of stored objects
-
-const INPUT_TYPES = ["text", "text", "email", "tel"];		
-
-const INPUT_PATTERNS = ["^[A-Z][a-z]+([ ][A-Z][a-z]+)*$",			// Words
-					  	"^[A-Z][a-z]+([ ][A-Z][a-z]+)*$",			// Words  
-					  	"[a-z0-9._%+-]+@[a-z0-9.-]+.[a-z]{2,}$",	// Email
-					  	"^[0-9]{1,3}-[0-9]{3}-[0-9]{3}-[0-9]{4}$"];	// Telephone 
+/* Other */
+import { INPUT_NAMES, INPUT_TYPES, INPUT_PATTERNS, INPUT_PLACEHOLDER } from "./constants"
  
-const INPUT_PLACEHOLDER = [	"John Jim", 
-							"Doe Poe",
-							"user@host.domain", 
-							"1-100-200-3000" ];
- 
-//==== App
+//==== App ====
+
 function App() {
-	// state
+	//--- I. State ---
+	
+	/* variables */
 	const [dataBase, setDataBase] = useState( [] );
 	const [inFocus, setInFocus] = useState(0);
 	const [inputs, setInputs] = useState( {} );
-  		// refs
+  	
+  	/* Refs */
 	const dbRef = collection( db, "data" );
  
-  	// display data
+ 	//--- II. Functions ---
+ 
+  	// 1. Display data
 	const resetForm = () => {
 		let _inputs = {};
 		for( const key in inputs ) {
@@ -43,18 +41,18 @@ function App() {
   			let array = data.docs.map( (doc) => ({id: doc.id, ...doc.data()}) );
   			// api calls
   			setDataBase( array );
-  			console.log("updated data");
+  			console.log("updated output");
 		};
 		getData();
 	};
 
-		// load on startup
-	useEffect( () => { loadData() }, [] );
+	/* Load on startup */
+	// eslint-disable-next-line
+	useEffect( () => { loadData() }, [] ); 
  	
- 	// create data	
+ 	// 2. Create data	
 	const handleSubmit = (event) => {
 		event.preventDefault();		
-		// load data
 		try {
     		addDoc(dbRef, inputs );
 		} catch(error) {
@@ -65,32 +63,39 @@ function App() {
 		resetForm();
 	} 
 	
+		// update inputs
 	const handleChange = (event) => {
 		const name = event.target.name;
 		const value = event.target.value;
 		setInputs( curr_inputs => ({...curr_inputs, [name]: value}) );
 	};
 
-	// delete data
+	// 2. Delete data
 	const handleDelete = async () => {
-		let id = dataBase[inFocus].id;
-		await deleteDoc( doc(db, "data", id) ); 
-		loadData();
-		
-		// change focus
-		let length = dataBase.length - 1;
-		if( inFocus === length && length !== 0 ) {
-			handleGet(length - 1);
+		let length = dataBase.length;
+		if( length > 0 ) {
+			let id = dataBase[inFocus].id;
+			await deleteDoc( doc(db, "data", id) ); 
+			loadData();
+			
+			// change focus
+			length -= 1;
+			if( inFocus === length && length !== 0 ) {
+				handleGet(length - 1);
+			}
+			resetForm(); 
 		}
-		resetForm(); 
 	};	 
 	
-	// update data
-	const handleEdit = async (index) => {		
-		let id = dataBase[inFocus].id;
-  		await updateDoc( doc(db, "data", id), inputs );
-  		loadData();
-  		resetForm();
+	// 3. Update data
+	const handleEdit = async (index) => {	
+		let length = dataBase.length;
+		if( length > 0 ) {				
+			let id = dataBase[inFocus].id;
+  			await updateDoc( doc(db, "data", id), inputs );
+  			loadData();
+  			resetForm();
+  		}
 	};
 	  	  
 	const handleGet = (index) => {
@@ -105,63 +110,83 @@ function App() {
 		setInFocus(index);
 	};
 	
-	const styleGet = (index) => {
-		return index === inFocus ? "buttonFocus" : "";
+	const styleGet = (index, style) => {
+		return index === inFocus ? style : "";
 	}
 	
-	// === JSX ===
+	//--- III. JSX ----
+	
 	return (
 	<div className="App">
-		{/* Form */}
-		<div>
-			<h1>CRUD application</h1>
-    		{/* inputs */}
-    		<form onSubmit={handleSubmit}>
-				{ INPUT_NAMES.map( (key, index) => { 
-					return (
-    				<label key={key}> 
-    					<span>{key}</span>
-						<input 
-							pattern={ INPUT_PATTERNS[index] }
-							placeholder={ INPUT_PLACEHOLDER[index] }
-							type={ INPUT_TYPES[index] } 
-							name={key} 
-							value={inputs[key] || ""} 
-							onChange={handleChange}
-							required
-						/>  
-					</label> ) 					
-				})}
-			
-    			{/* create */}
-    			<button type="submit"> Create </button>
+		<div className="mainBox">
+			<h1>Contact Information</h1>
+			{/* Form */}
+    		<form className="inputForm" id="input_fields" onSubmit={handleSubmit}>
+    			<div className="inputList">
+					{ INPUT_NAMES.map( (key, index) => { 
+						return (
+    					<label key={key}> 
+    						<span>{key}</span>
+							<input 
+								pattern={ INPUT_PATTERNS[index] }
+								placeholder={ INPUT_PLACEHOLDER[index] }
+								type={ INPUT_TYPES[index] } 
+								name={key} 
+								value={inputs[key] || ""} 
+								onChange={handleChange}
+								min="1"
+								required
+							/>  
+						</label> ) 					
+					})}
+				</div>
     		</form>
-		</div>
-		
-		{/* delete and update */}
-    	<button onClick={handleDelete}> Delete </button>	
-  		<button onClick={handleEdit}> Update </button>
-				
-		{/* list output */}
-		<div>
-			{ dataBase.map( (data, index) => {
-				// properties
-				const outputList = INPUT_NAMES.map( name => {
-					return ( <span key={name}> { data[name] } </span> )	
-				});
-				
-				// display
-				const id = data.id; 
-				
-				return (
-				<div key={id} className={styleGet(index)}>
-  					{ outputList }
-  					
-  					<button onClick={ () => {handleGet(index)} }> 
-  						Select
-  					</button>
-  				</div>); 
-  			})}
+			
+			{/* create, delete and update buttons */}
+			<div className="submitButtons">
+				<button className="create" type="submit" form="input_fields"> Create </button>	
+    			<button className="delete" onClick={handleDelete}> Delete </button>	
+  				<button className="" onClick={handleEdit}> Update </button>
+    		</div>				
+			{/* list output */}
+			
+			<table className="tableList" cellspacing="0" cellpadding="7">
+				{/* header */}
+				<thead>
+				{ dataBase.length === 0 ? (<></>) : (
+					<tr>
+    					{ INPUT_NAMES.map( name => {
+							return ( <th key={name}> {name} </th> )	
+						})}
+  					</tr>
+  				)}
+  				</thead>
+  			
+  				<tbody>
+  				{/* content */}
+				{ dataBase.map( (data, i) => {
+					// object properties
+					const outputList = INPUT_NAMES.map( name => {
+						return ( <td key={name} className={ styleGet(i, "focusText") }> 
+									{ data[name] } 
+								</td> )	
+					});
+					
+					// display data
+					const id = data.id; 
+					return (
+					<tr key={id}>
+  						{ outputList }
+  						<td>
+  							<button className={ styleGet(i, "focusBtn") } 
+  									onClick={ () => {handleGet(i)} }> 
+  								Select
+  							</button>
+  						</td>
+  					</tr>); 
+  				})}
+  				</tbody>
+			</table>
 		</div>    	
 	</div>
 	);
